requirements.txt
Flask==2.3.3
flask_pymongo==2.3.0
pymongo==4.7.1
python-dotenv==1.1.1
